Added Exception managment

+++ characters/src/main/java/dev/tony/characters/RestApi/CharactersServiceIMPL.java
@@ -1,7 +1,10 @@
 package dev.tony.characters.RestApi;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.crossstore.ChangeSetPersister.NotFoundException;
 import org.springframework.stereotype.Service;
+
+import io.micrometer.common.util.StringUtils;
 
 import java.util.List;
 import java.util.Optional;
@@ -16,7 +19,10 @@ public class CharactersServiceIMPL {
         return charactersRepository.findAll();
     }
 
-    public Characters addCharacter(Characters character) {
+    public Characters addCharacter(Characters character) throws NotFoundException{
+        if (character == null || StringUtils.isBlank(character.getAlias())) {
+            throw new NotFoundException();
+        }        
         return charactersRepository.save(character);
     }
 
@@ -24,11 +30,21 @@ public class CharactersServiceIMPL {
         return charactersRepository.save(character);
     }
 
-    public Optional<Characters> findCharacterById(String id) {
+    public Optional<Characters> findCharacterById(String id) throws NotFoundException {
+        if (StringUtils.isBlank(id)) {
+            throw new NotFoundException();
+        }
+        Optional<Characters> character = charactersRepository.findById(id);
+        if (!character.isPresent()) {
+            throw new NotFoundException();
+        }
         return charactersRepository.findById(id);
     }
 
-    public void deleteCharacterById(String id) {
+    public void deleteCharacterById(String id) throws NotFoundException {
+        if (StringUtils.isBlank(id)) {
+            throw new NotFoundException();
+        }
         charactersRepository.deleteById(id);
     }
